feat: introduced data polling for map view and refactored whole view, dividing it on couple smaller components
