+++ nerdlets/geo-ops-nerdlet/styles.scss
@@ -11,7 +11,7 @@
 // Main (UI/View) pages
 @import './CreateMap.scss';
 @import './MapList.scss';
-@import './ViewMap.scss';
+@import './ViewMap/ViewMap.scss';
 @import './DefineLocations.scss';
 @import './MapLocationData.scss';
 @import './EmptyState.scss';
